modify signatures of gsva() generic and methods for new API
for the new API, the signature for gsva() generic and methods on which
S4 dispatch is done is just the parameter object, i.e. (param, ...).
(for the old API, this used to be (expr, gset.idx.list, ...).)
This change is made in code and docs (plus a few minor fixes to the
documentation, mostly cosmetic).

however, I haven't found a way to make these changes AND at the same
time retain the now defunct code and docs of old API; at least not
without 'R CMD check' detecting and complaining about it.